// src/core/models/ModuleField.ts
export interface ModuleField {
    Color: string;
    CtrlType: string;
    DefaultValue: string;
    FldLabel: string;
    FldName: string;
    FldType: string;
    ListSource: string;
    Required: string;
    Validator: string;
    GroupOnLoad: string;
    ReadOnly: string;
    PinOnLoad: string;
    CallBack: string;
}